+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.7.0] - 2019-09-09
+### Added
+- Support custom types in ObjectView
+- Add getValue methods to View classes
+- Add toJSON methods and deprecate toObject in View classes 
+
+## [1.6.1] - 2019-08-20
+### Fixed
+- Add iterators in TypeScript definitions
+
 ## [1.6.0] - 2019-08-17
 ### Added
 - Add CollectionView